refactor code to break down hooks and remove redundant code
