+++ apps/web/src/components/utils/hooks/useMessagePreviews.tsx
@@ -3,12 +3,11 @@ import buildConversationId from '@lib/buildConversationId';
 import chunkArray from '@lib/chunkArray';
 import { buildConversationKey, parseConversationKey } from '@lib/conversationKey';
 import conversationMatchesProfile from '@lib/conversationMatchesProfile';
-import type { Conversation, Stream } from '@xmtp/xmtp-js';
+import type { Conversation } from '@xmtp/xmtp-js';
 import { SortDirection } from '@xmtp/xmtp-js';
 import type { DecodedMessage } from '@xmtp/xmtp-js/dist/types/src/Message';
 import type { Profile } from 'lens';
 import { useProfilesLazyQuery } from 'lens';
-import { useRouter } from 'next/router';
 import { useEffect, useState } from 'react';
 import { useAppStore } from 'src/store/app';
 import { useMessageStore } from 'src/store/message';
@@ -16,27 +15,21 @@ import { useMessageStore } from 'src/store/message';
 const MAX_PROFILES_PER_REQUEST = 50;
 
 const useMessagePreviews = () => {
-  const router = useRouter();
   const currentProfile = useAppStore((state) => state.currentProfile);
   const conversations = useMessageStore((state) => state.conversations);
-  const setConversations = useMessageStore((state) => state.setConversations);
   const messageProfiles = useMessageStore((state) => state.messageProfiles);
   const setMessageProfiles = useMessageStore((state) => state.setMessageProfiles);
   const previewMessages = useMessageStore((state) => state.previewMessages);
   const setPreviewMessages = useMessageStore((state) => state.setPreviewMessages);
   const selectedProfileId = useMessageStore((state) => state.selectedProfileId);
-  const setSelectedProfileId = useMessageStore((state) => state.setSelectedProfileId);
   const setPreviewMessage = useMessageStore((state) => state.setPreviewMessage);
-  const reset = useMessageStore((state) => state.reset);
-  const { client, loading: creatingXmtpClient } = useXmtpClient();
+  const { client } = useXmtpClient();
   const [profileIds, setProfileIds] = useState<Set<string>>(new Set<string>());
   const [messagesLoading, setMessagesLoading] = useState<boolean>(true);
   const [profilesLoading, setProfilesLoading] = useState<boolean>(false);
   const [profilesError, setProfilesError] = useState<Error | undefined>();
   const [loadProfiles] = useProfilesLazyQuery();
-  const selectedTab = useMessageStore((state) => state.selectedTab);
-  const [profilesToShow, setProfilesToShow] = useState<Map<string, Profile>>(new Map());
-  const [requestedCount, setRequestedCount] = useState(0);
+  const addMessages = useMessageStore((state) => state.addMessages);
 
   const getProfileFromKey = (key: string): string | null => {
     const parsed = parseConversationKey(key);
@@ -100,7 +93,6 @@ const useMessagePreviews = () => {
     }
     const matcherRegex = conversationMatchesProfile(currentProfile.id);
     let messageStream: AsyncGenerator<DecodedMessage>;
-    let conversationStream: Stream<Conversation>;
 
     const streamAllMessages = async () => {
       messageStream = await client.conversations.streamAllMessages();
@@ -110,6 +102,7 @@ const useMessagePreviews = () => {
         if (conversationId && matcherRegex.test(conversationId)) {
           const key = buildConversationKey(message.conversation.peerAddress, conversationId);
           setPreviewMessage(key, message);
+          addMessages(key, [message]);
         }
       }
     };
@@ -129,22 +122,13 @@ const useMessagePreviews = () => {
       return { key, message: newMessages[0] };
     };
 
-    const listConversations = async () => {
+    const loadPreviewMessages = async () => {
       setMessagesLoading(true);
       const newPreviewMessages = new Map(previewMessages);
-      const newConversations = new Map(conversations);
+      const newConversations = Array.from(conversations.values());
       const newProfileIds = new Set(profileIds);
-      const convos = await client.conversations.list();
-      const matchingConvos = convos.filter(
-        (convo) => convo.context?.conversationId && matcherRegex.test(convo.context.conversationId)
-      );
 
-      for (const convo of matchingConvos) {
-        const key = buildConversationKey(convo.peerAddress, convo.context?.conversationId as string);
-        newConversations.set(key, convo);
-      }
-
-      const previews = await Promise.all(matchingConvos.map(fetchMostRecentMessage));
+      const previews = await Promise.all(newConversations.map(fetchMostRecentMessage));
 
       for (const preview of previews) {
         const profileId = getProfileFromKey(preview.key);
@@ -156,18 +140,10 @@ const useMessagePreviews = () => {
         }
       }
       setPreviewMessages(newPreviewMessages);
-      setConversations(newConversations);
       setMessagesLoading(false);
       if (newProfileIds.size > profileIds.size) {
         setProfileIds(newProfileIds);
       }
-    };
-
-    const closeConversationStream = async () => {
-      if (!conversationStream) {
-        return;
-      }
-      await conversationStream.return();
     };
 
     const closeMessageStream = async () => {
@@ -176,76 +152,18 @@ const useMessagePreviews = () => {
       }
     };
 
-    const streamConversations = async () => {
-      closeConversationStream();
-      conversationStream = (await client?.conversations?.stream()) || [];
-      const matcherRegex = conversationMatchesProfile(currentProfile?.id);
-      for await (const convo of conversationStream) {
-        // Ignore any new conversations not matching the current profile
-        if (!convo.context?.conversationId || !matcherRegex.test(convo.context.conversationId)) {
-          continue;
-        }
-        const newConversations = new Map(conversations);
-        const newProfileIds = new Set(profileIds);
-        const key = buildConversationKey(convo.peerAddress, convo.context.conversationId);
-        newConversations.set(key, convo);
-        const profileId = getProfileFromKey(key);
-        if (profileId && !profileIds.has(profileId)) {
-          newProfileIds.add(profileId);
-          setProfileIds(newProfileIds);
-        }
-        setConversations(newConversations);
-      }
-    };
-
-    listConversations();
-    streamConversations();
+    loadPreviewMessages();
     streamAllMessages();
 
     return () => {
-      closeConversationStream();
       closeMessageStream();
     };
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [client, currentProfile?.id, selectedProfileId]);
-
-  useEffect(() => {
-    if (selectedProfileId && currentProfile?.id !== selectedProfileId) {
-      reset();
-      setSelectedProfileId(currentProfile?.id);
-      router.push('/messages');
-    } else {
-      setSelectedProfileId(currentProfile?.id);
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [currentProfile]);
-
-  useEffect(() => {
-    const partitionedProfiles = Array.from(messageProfiles).reduce(
-      (result, [key, profile]) => {
-        const message = previewMessages.get(key);
-        if (message) {
-          if (profile.isFollowedByMe) {
-            result[0].set(key, profile);
-          } else {
-            result[1].set(key, profile);
-          }
-        }
-        return result;
-      },
-      [new Map<string, Profile>(), new Map<string, Profile>()]
-    );
-    setProfilesToShow(selectedTab === 'Following' ? partitionedProfiles[0] : partitionedProfiles[1]);
-    setRequestedCount(partitionedProfiles[1].size);
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [previewMessages, messageProfiles, selectedTab]);
+  }, [client, currentProfile?.id, selectedProfileId, conversations]);
 
   return {
-    authenticating: creatingXmtpClient,
-    loading: messagesLoading || profilesLoading,
+    previewLoading: messagesLoading || profilesLoading,
     messages: previewMessages,
-    profilesToShow,
-    requestedCount,
     profilesError: profilesError
   };
 };
